Classes to manage ephemeral test processes
The current test suite starts salt-master, salt-minion and other processes at
the start and expects the same processes to continue running throughout the
entire test suite.  This is a problem when an integration test needs to test
starting/stopping/restarting the same process.  These are classes for managing
short-lived processes for the test suite.